009-09-18    <dmg@uvic.ca>
Patch submitted by Kornel Benko

* ptpicker.c: Casted some values in printf to avoid compiler
warnings.

* CMakeLists.txt: As allways one has to call cmake
-DCPACK_BINARY_DEB:BOOL=ON to create a package (debian/ubuntu in
this example)

        make package creates then the package libpano13-2.9.15-Linux.deb.
           (Values automatically taken from configure.ac) The package will
install the header-files and the (shared) libpano13 library.

+++ ptpicker.c
@@ -731,7 +731,7 @@ int writeProject( AlignInfo *g, fullPath *pFile)
 		
 		
 	
-	sprintf(line, "p f%d w%ld h%ld v%lg u%d %s n\"%s\" %s\n\n", pf, g->pano.width, g->pano.height, g->pano.hfov, g->st.feather, cv, g->pano.name, ch );
+	sprintf(line, "p f%d w%ld h%ld v%lg u%d %s n\"%s\" %s\n\n", pf, (long int) g->pano.width, (long int) g->pano.height, g->pano.hfov, g->st.feather, cv, g->pano.name, ch );
 	count = strlen( line ); mywrite( fnum, count, line );
 	
 	for(i=0; i<g->numIm; i++)
@@ -743,7 +743,7 @@ int writeProject( AlignInfo *g, fullPath *pFile)
 		else
 			sprintf(ch, "v%lg", g->im[i].hfov);
 			
-		sprintf(line, "i f%ld w%ld h%ld y%lg p%lg r%lg %s %s n\"%s\" ", im->format, im->width, im->height, 
+		sprintf(line, "i f%ld w%ld h%ld y%lg p%lg r%lg %s %s n\"%s\" ", (long int) im->format, (long int) im->width, (long int) im->height, 
 											im->yaw, im->pitch, im->roll, 
 											(im->cP.correction_mode & correction_mode_morph ? "o" : "" ),
 											ch, im->name );
@@ -784,8 +784,8 @@ int writeProject( AlignInfo *g, fullPath *pFile)
 		}
 
 		if( g->im[i].selection.bottom != 0 || g->im[i].selection.right != 0 ){
-			sprintf( ch, " S%ld,%ld,%ld,%ld ",g->im[i].selection.left, g->im[i].selection.right,
-						       g->im[i].selection.top, g->im[i].selection.bottom );
+			sprintf( ch, " S%ld,%ld,%ld,%ld ",(long int) g->im[i].selection.left, (long int) g->im[i].selection.right,
+						       (long int) g->im[i].selection.top, (long int) g->im[i].selection.bottom );
 			strcat(line, ch);
 		}
